Changed texture and submap generation to modify pixels

+++ TerrainGenerator/HeightMap.h
@@ -7,16 +7,16 @@ class HeightMap {
         int width = 0;
         int height = 0;
 
-        double* values;
+        float* values;
 
     public:
         HeightMap(int width, int height);
-        HeightMap(int width, int height, double* heights);
+        HeightMap(int width, int height, float* heights);
         ~HeightMap();
 
-        double* getValues();
-        void setHeightAt(int x, int y, double value);
-        double getHeightAt(int x, int y);
+        float* getValues();
+        void setHeightAt(int x, int y, float value);
+        float getHeightAt(int x, int y);
 
         int getMapWidth();
         int getMapHeight();
